+++ src/index.js
@@ -1,2 +1,3 @@
 export { Transport } from './transport';
-export Socket from './socket';
+export { Socket } from './socket';
+export Bridge from './bridge';